New site based on gatsby.
I took the simplest template I could find and adapted it to our needs. Because there
was no way that developers were not going to complain loudly if I point them to a
webpage that is an image. I still plan to keep things simple and just integrate
netlify and intercom and some form builder